requirements.txt
python-dotenv
langchain
openai
youtube-transcript-api
faiss-cpu
streamlit  
tiktoken 
